requirements.txt
aiogram==3.4.1
aiohttp==3.9.3
aiosqlite==0.19.0
python-dotenv==1.0.1
openai==1.14.0
cryptography==42.0.5